# [baremetal] Add SCA countermeasures to AES