## Description
Add SCA countermeasures to AES implementation. Add total of 5 dummy calculation rounds to AES encryption/decryption.
 * 1 additional dummy round calculated with first and last round (order randomised).
 * 3 additional dummy AES rounds calculated with middle rounds (order randomized)

__Impact on code-size:__

| | GCC | ARMC5 | ARMC6 |
| --- | --- | --- | --- |
| `libmbedtls.a` before | 15069 | 16511| 17065|
| `libmbedtls.a` after | 15069 | 16511| 17065|
| `libmbedcrypto.a` before | 17999 | 18427| 20599|
| `libmbedcrypto.a` after | 17843 | 18129| 20477 |
| `libmbedx509.a` before | 4729 | 4868| 5159|
| `libmbedx509.a` after | 4729 | 4868 | 5159|
| gain in bytes TLS | == | == | == |
| gain in bytes Crypto | 156 | 298 | 122 |
| gain in bytes X.509 | == | == | == |
| gain Total | 156 | 298 | 122 |

__Impact on performance (countermeasures activated):__

|      | before | after |
| --- | --- | --- |
|MD4                      |       3867 KB/s  |      3846 KB/s|
|MD5                      |       3134 KB/s  |      2873 KB/s|
|RIPEMD160            |        730 KB/s  |       725 KB/s|
|SHA-1                    |       1262 KB/s  |      1259 KB/s|
|SHA-256                  |        621 KB/s  |       619 KB/s|
|SHA-512                  |        383 KB/s  |       384 KB/s|
|ARC4                     |       2783 KB/s  |      2781 KB/s|
|3DES                     |        140 KB/s  |       137 KB/s|
|DES                      |        414 KB/s  |       412 KB/s|
|3DES-CMAC                |        128 KB/s  |       126 KB/s|
|AES-CBC-128              |        536 KB/s  |       258 KB/s|
|AES-CBC-192              |        472 KB/s  |       234 KB/s|
|AES-CBC-256              |        422 KB/s  |       213 KB/s|
|AES-CTR-128              |        563 KB/s  |       265 KB/s|
|AES-CTR-192              |        493 KB/s  |       240 KB/s|
|AES-CTR-256              |        437 KB/s  |       217 KB/s|
|AES-GCM-128              |        318 KB/s  |       193 KB/s|
|AES-GCM-192              |        293 KB/s  |       179 KB/s|
|AES-GCM-256              |        272 KB/s  |       166 KB/s|
|AES-CCM-128              |        240 KB/s  |       122 KB/s|
|AES-CCM-192              |        213 KB/s  |       112 KB/s|
|AES-CCM-256              |        192 KB/s  |       102 KB/s|
|AES-CMAC-128             |        471 KB/s  |       243 KB/s|
|AES-CMAC-192             |        419 KB/s  |       221 KB/s|
|AES-CMAC-256             |        378 KB/s  |       202 KB/s|
|AES-CMAC-PRF-128         |        471 KB/s  |       243 KB/s|
|CAMELLIA-CBC-128         |        489 KB/s  |       489 KB/s|
|CAMELLIA-CBC-192         |        381 KB/s  |       381 KB/s|
|CAMELLIA-CBC-256         |        381 KB/s  |       381 KB/s|
|BLOWFISH-CBC-128         |        692 KB/s  |       692 KB/s|
|BLOWFISH-CBC-192         |        692 KB/s  |       692 KB/s|
|BLOWFISH-CBC-256         |        692 KB/s  |       692 KB/s|
|CTR_DRBG (NOPR)          |        419 KB/s  |       211 KB/s|
|CTR_DRBG (PR)            |        311 KB/s  |       159 KB/s|
|HMAC_DRBG SHA-1 (NOPR)   |         90 KB/s  |        90 KB/s|
|HMAC_DRBG SHA-1 (PR)     |         84 KB/s  |        84 KB/s|
|HMAC_DRBG SHA-256 (NOPR) |         65 KB/s  |        65 KB/s|
|HMAC_DRBG SHA-256 (PR)   |         65 KB/s  |        65 KB/s|
|RSA-2048                 |      63 ms/ public  |     64 ms/ public|
|RSA-2048                 |    2448 ms/private  |   2455 ms/private|
|ECDSA-secp384r1          |     940 ms/sign  |    950 ms/sign|
|ECDSA-secp256r1          |     628 ms/sign  |    636 ms/sign|
|ECDSA-secp384r1          |    1825 ms/verify  |   1848 ms/verify|
|ECDSA-secp256r1          |    1242 ms/verify  |   1244 ms/verify|
|ECDHE-secp384r1          |    1762 ms/handshake  |   1777 ms/handshake|
|ECDHE-secp256r1          |    1183 ms/handshake  |   1209 ms/handshake|
|ECDH-secp384r1           |     855 ms/handshake  |    864 ms/handshake|
|ECDH-secp256r1           |     582 ms/handshake  |    588 ms/handshake|
|ECDHE-Curve25519         |     981 ms/handshake  |   1007 ms/handshake|
|ECDH-Curve25519          |     495 ms/handshake  |    509 ms/handshake|

## Status
**READY**

## Additional comments
~Work in progress, not ready yet.~

This PR is made on top of PR ~#2870~~#2891~ baremetal