@mpg 2019-10-23
@artokin I'm surprised by the presence of a few commits about X.509 here (from "Changed x509.c to be one single compilation unit for all x509 files." to "Changed generate_visualc_files.pl to exclude x509_xxx files"). Is that intentional, or just an unintended side-effect of how you prepared the branch?