@artokin 2019-10-24
Thanks for your comments @mpg , I'll address them next.
I just rebased this PR on top of updated #2891