@artokin 2019-12-10
Rebased on top of 0252cd6.