@mpg 2019-12-18
Removing "needs: preceding PR" as all prerequisites are now merged.